# Dryft/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Drift API
  description: Backend API for the Drift social/dating application.
  version: 1.0.0
  contact:
    name: Drift Team
servers:
  - url: https://api.drift.app
    description: Production
  - url: http://localhost:8080
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /v1/creators/{creatorID}/items:
    get:
      tags: [Creators]
      summary: Get a creator's published items
      security:
        - {}
        - BearerAuth: []
      operationId: getCreatorItems
      parameters:
        - name: creatorID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Creator's items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/StoreItem"
                  limit:
                    type: integer
                  offset:
                    type: integer
...
  /v1/creators/items:
    get:
      tags: [Creators]
      summary: Get own items as a creator
      operationId: getMyItems
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ItemStatus"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Creator's own items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/StoreItem"
                  limit:
                    type: integer
                  offset:
                    type: integer

  # ---------------------------------------------------------------------------
  # Webhooks
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
...
  schemas:
...
    StoreItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        creator_id:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/ItemType"
        name:
          type: string
        description:
          type: string
        price:
          type: integer
          description: Price in cents
        currency:
          type: string
          default: usd
        thumbnail_url:
          type: string
        preview_url:
          type: string
        asset_bundle:
          type: object
        tags:
          type: array
          items:
            type: string
        purchase_count:
          type: integer
        rating:
          type: number
          format: float
        is_featured:
          type: boolean
        status:
          $ref: "#/components/schemas/ItemStatus"
        created_at:
          type: string
          format: date-time
    ItemType:
      type: string
      enum: [avatar, outfit, toy, effect, gesture]
    ItemStatus:
      type: string
      enum: [draft, pending, approved, rejected, disabled]
...